requirements.txt
requests
python-dotenv
pandas
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pickle5
